Moved a bunch of functions from pcheckRegistration to their own modules
